Reorganize "library" code and stores
