+++ .eslintignore
@@ -1,1 +1,1 @@
-src/tonejs-piano.js
+src/lib/tonejs-piano.js
